Get to actually compile

+++ pwm.c
@@ -1,5 +1,6 @@
+#include <xc.h>
 #include "pwm.h"
 
-void set_motor_pwm(uint16_t duty) {
+void set_motor_pwm(uint8_t duty) {
     // pwm_set_duty_cycle(duty);
 }
